Display username of trip creator
Translate the User Google ID of the trip creator to his
username by calling get user from the server.
we might want to replace the logic in the future, and instead,
have a Map structe to store IDs and Usernames, so we won't need
to send api call for each trip, alternatvly we can Include the username
in the returned trip response.